+++ cfbd_json_py/rankings.py
@@ -1,3 +1,9 @@
+# Creation Date: 08/30/2023 01:13 EDT
+# Last Updated Date: 10/06/2023 07:53 PM EDT
+# Author: Joseph Armstrong (armstrongjoseph08@gmail.com)
+# File Name: rankings.py
+# Purpose: Houses functions pertaining to CFB poll data within the CFBD API.
+####################################################################################################
 
 def get_cfbd_poll_rankings(
         season: int,