+++ cfbd_json_py/ratings.py
@@ -1,8 +1,14 @@
+# Creation Date: 08/30/2023 01:13 EDT
+# Last Updated Date: 10/06/2023 07:53 PM EDT
+# Author: Joseph Armstrong (armstrongjoseph08@gmail.com)
+# File Name: ratings.py
+# Purpose: Houses functions pertaining to CFB team rating data within the CFBD API.
+####################################################################################################
 
 def get_cfbd_sp_plus_ratings(
         api_key: str = None,
         api_key_dir: str = None,
-        year: int = None,
+        season: int = None,
         team: int = None,
         # Either `year` or `team` have to be not null for this function to work.
 
@@ -18,7 +24,7 @@ def get_cfbd_sp_plus_ratings(
 def get_cfbd_srs_ratings(
         api_key: str = None,
         api_key_dir: str = None,
-        year: int = None,
+        season: int = None,
         team: int = None,
         # Either `year` or `team` have to be not null for this function to work.
         conferenece: str = None,