+++ cfbd_json_py/teams.py
@@ -1,3 +1,10 @@
+# Creation Date: 08/30/2023 01:13 EDT
+# Last Updated Date: 10/06/2023 07:54 PM EDT
+# Author: Joseph Armstrong (armstrongjoseph08@gmail.com)
+# File Name: teams.py
+# Purpose: Houses functions pertaining to CFB team data within the CFBD API.
+####################################################################################################
+
 def get_cfbd_team_information(
         api_key: str = None,
         api_key_dir: str = None,
@@ -15,7 +22,7 @@ def get_cfbd_team_information(
 def get_cfbd_fbs_team_list(
         api_key: str = None,
         api_key_dir: str = None,
-        year: int = None,
+        season: int = None,
 
         return_as_dict: bool = False):
     """
@@ -30,7 +37,7 @@ def get_cfbd_team_rosters(
         api_key: str = None,
         api_key_dir: str = None,
         team: str = None,
-        year: int = None,
+        season: int = None,
 
         return_as_dict: bool = False):
     """
@@ -44,7 +51,7 @@ def get_cfbd_team_rosters(
 def get_cfbd_team_talent_rankings(
         api_key: str = None,
         api_key_dir: str = None,
-        year: int = None,
+        season: int = None,
 
         return_as_dict: bool = False):
     """
@@ -60,8 +67,8 @@ def get_cfbd_team_matchup_history(
         team_2: str,
         api_key: str = None,
         api_key_dir: str = None,
-        min_year: int = None,
-        max_year: int = None,
+        min_season: int = None,
+        max_season: int = None,
 
         return_as_dict: bool = False):
     """